requirements-gpu.txt
opencv-python
onnxruntime-gpu[cuda,cudnn]